Add files via upload
This simple Python script will check for a specific service and send an alert to the appropriate people when the service goes down and will attempt to restart it. You will have to fill in your own information for either one specific service or it can be modified to include a number of services to monitor or watch and respond appropriately. This is just the infrastructure to many more things possible!